Split main.py into inference and traces modules
Extract inference logic (create_language_model, run_baseline, run_its,
calculate_cost, detect_question_type) into inference.py (~370 lines)
and trace building (build_trace, _build_pf_trace) into traces.py
(~170 lines). main.py is now focused on app setup and route handlers
(~400 lines, down from ~1270).

All 116 passing tests continue to pass with updated imports.

Closes #47
