+++ demo_ui/backend/main.py
@@ -1,26 +1,18 @@
 """
 FastAPI backend for ITS demo.
 
-Exposes two endpoints:
-- GET /health: Health check
-- POST /compare: Compare baseline vs ITS
+App setup, CORS, static file serving, and route handlers.
+Inference logic is in inference.py, trace building in traces.py.
 """
 
 import asyncio
-import json
 import logging
 import os
-import re
 import socket
-import time
-import traceback
 import uuid
 from pathlib import Path
-from typing import Dict
 from urllib.parse import urlparse
 
-import litellm
-import numpy as np
 from dotenv import load_dotenv
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
@@ -48,131 +40,9 @@ else:
     logger.warning("OPENAI_API_KEY not found in environment!")
 
 from backend.evaluation import evaluate_correctness
-from its_hub.lms import OpenAICompatibleLanguageModel, LiteLLMLanguageModel, StepGeneration
-from its_hub.algorithms import (
-    BestOfN,
-    SelfConsistency,
-    BeamSearch,
-    ParticleFiltering,
-    EntropicParticleFiltering,
-    ParticleGibbs,
-)
-try:
-    from its_hub.integration.reward_hub import LLMJudgeRewardModel
-except ImportError:
-    LLMJudgeRewardModel = None
-from its_hub.types import ChatMessage
-from its_hub.utils import extract_content_from_lm_response, QWEN_SYSTEM_PROMPT
-from its_hub.base import AbstractLanguageModel
-from its_hub.algorithms.self_consistency import create_regex_projection_function
+from its_hub.utils import QWEN_SYSTEM_PROMPT
 
-from .config import get_model_config, get_api_key, MODEL_REGISTRY, ModelConfig
-from .vertex_lm import VertexAIClaudeModel, VertexAIGeminiModel
-from .llm_prm import LLMProcessRewardModel
-
-# ── Algorithm / model defaults ────────────────────────────────────────
-DEFAULT_JUDGE_MODEL = "gpt-4.1-mini"
-DEFAULT_PRM_MODEL = "gpt-4.1-mini"
-DEFAULT_STEP_GEN_MAX_STEPS = 8
-DEFAULT_STEP_GEN_TEMPERATURE = 0.8
-DEFAULT_STEP_GEN_TOKEN = "\n\n"
-DEFAULT_PRM_TEMPERATURE = 0.3
-BEAM_WIDTH_MIN = 2
-BEAM_WIDTH_MAX = 4
-GIBBS_ITERATIONS_MIN = 2
-GIBBS_ITERATIONS_MAX = 3
-GIBBS_ITERATIONS_DIVISOR = 4
-
-
-def parse_tool_args(tool_args):
-    """Parse tool arguments, handling JSON string encoding."""
-    if isinstance(tool_args, str):
-        try:
-            return json.loads(tool_args)
-        except json.JSONDecodeError:
-            return {}
-    return tool_args
-
-
-def calculate_cost(
-    model_config: ModelConfig,
-    input_tokens: int,
-    output_tokens: int
-) -> float:
-    """
-    Calculate cost in USD based on token usage and model pricing.
-
-    Returns:
-        Cost in USD (rounded to 4 decimal places)
-    """
-    input_cost_per_1m = model_config.get("input_cost_per_1m", 0.0)
-    output_cost_per_1m = model_config.get("output_cost_per_1m", 0.0)
-
-    if input_cost_per_1m == 0.0 and output_cost_per_1m == 0.0:
-        return 0.0
-
-    input_cost = (input_tokens / 1_000_000) * input_cost_per_1m
-    output_cost = (output_tokens / 1_000_000) * output_cost_per_1m
-    total_cost = input_cost + output_cost
-
-    return round(total_cost, 6)  # Round to 6 decimal places for precision
-
-
-def detect_question_type(
-    question: str,
-    enable_tools: bool = False,
-    question_metadata: dict | None = None
-) -> str:
-    """
-    Detect question type: 'math', 'tool_calling', or 'general'.
-
-    Args:
-        question: The question text to analyze
-        enable_tools: Whether tools are enabled for this question
-        question_metadata: Optional metadata about the question
-
-    Returns:
-        Question type: "math", "tool_calling", or "general"
-    """
-    # Check metadata first
-    if question_metadata:
-        if question_metadata.get("expected_tools") or question_metadata.get("source") == "tool_calling":
-            return "tool_calling"
-
-    # Check if tools enabled
-    if enable_tools:
-        return "tool_calling"
-
-    # Check for math patterns using weighted scoring to reduce false positives.
-    # Strong indicators (LaTeX notation) are sufficient on their own.
-    # Weak indicators require at least 2 matches to trigger math detection.
-    strong_indicators = [
-        r'\\frac', r'\\boxed', r'\\sum', r'\\int', r'\\sqrt',
-        r'\\begin\{', r'\\mathbb', r'\\left', r'\\right',
-        r'\$[^$]+\$',  # Paired dollar signs (LaTeX inline math)
-    ]
-    weak_indicators = [
-        r'probability', r'find the value', r'what is the.*term',
-        r'solve for [a-z]', r'evaluate the (integral|derivative|limit)',
-        r'\bx\s*[\+\-\*\/\^]\s*\d', r'\d\s*[\+\-\*\/\^]\s*x',  # algebraic expressions
-    ]
-
-    for pattern in strong_indicators:
-        if re.search(pattern, question):
-            return "math"
-
-    weak_count = sum(1 for p in weak_indicators if re.search(p, question, re.IGNORECASE))
-    if weak_count >= 2:
-        return "math"
-
-    return "general"
-
-
-def create_math_projection_function():
-    """Create projection function for extracting boxed mathematical answers."""
-    return create_regex_projection_function(r'\\boxed\{([^}]+)\}')
-
-
+from .config import get_model_config, MODEL_REGISTRY
 from .example_questions import (
     get_all_questions,
     get_questions_by_algorithm,
@@ -184,22 +54,17 @@ from .models import (
     CompareResponse,
     ResultDetail,
     HealthResponse,
-    CandidateResponse,
-    SelfConsistencyTrace,
-    BestOfNTrace,
-    BeamSearchTrace,
-    ParticleFilteringTrace,
-    ParticleGibbsTrace,
-    ToolCall,
-    ToolVotingTrace,
 )
-from .tools import get_tool_schemas, execute_tool
-from its_hub.algorithms.self_consistency import SelfConsistencyResult
-from its_hub.algorithms.bon import BestOfNResult
-from its_hub.algorithms.beam_search import BeamSearchResult
-from its_hub.algorithms.particle_gibbs import ParticleFilteringResult, ParticleGibbsResult
+from .inference import (
+    create_language_model,
+    run_baseline,
+    run_its,
+    calculate_cost,
+    detect_question_type,
+)
 
-# Create FastAPI app
+# ── FastAPI App ───────────────────────────────────────────────────────
+
 app = FastAPI(
     title="ITS Demo API",
     description="Demo API for comparing baseline vs ITS (Inference-Time Scaling)",
@@ -219,6 +84,8 @@ app.add_middleware(
 
 frontend_dir = Path(__file__).parent.parent / "frontend"
 
+
+# ── Routes ────────────────────────────────────────────────────────────
 
 @app.get("/")
 async def serve_frontend():
@@ -357,8 +224,6 @@ async def list_models(use_case: str | None = None):
 
         if server_available:
             available_models.append(model_entry)
-        else:
-            logger.debug(f"Skipping model {model_id} - server at {base_url} not available")
 
     return {"models": available_models}
 
@@ -401,570 +266,6 @@ async def list_examples(algorithm: str | None = None, use_case: str | None = Non
             for q in questions
         ]
     }
-
-
-def create_language_model(
-    model_id: str,
-    system_prompt: str | None = None
-) -> AbstractLanguageModel:
-    """
-    Create a language model instance based on the model configuration.
-
-    All models use OpenAI-compatible endpoints except Vertex AI models.
-
-    Args:
-        model_id: Model identifier from the registry
-        system_prompt: Optional system prompt to prepend to all messages
-    """
-    model_config = get_model_config(model_id)
-    provider = model_config.get("provider", "openai")
-
-    if provider == "vertex_ai":
-        # Use native Vertex AI SDK for Claude and Gemini models (NOT OpenAI-compatible)
-        vertex_project = model_config.get("vertex_project")
-        vertex_location = model_config.get("vertex_location")
-
-        if not vertex_project or vertex_project == "your-gcp-project-id":
-            raise ValueError(
-                "VERTEX_PROJECT not configured. Please set VERTEX_PROJECT "
-                "environment variable in your .env file"
-            )
-
-        model_name = model_config["model_name"]
-
-        # Determine if it's a Claude or Gemini model based on model name
-        if "claude" in model_name.lower():
-            logger.info(
-                f"Creating Vertex AI Claude model: {model_name} "
-                f"(project: {vertex_project}, location: {vertex_location})"
-            )
-            return VertexAIClaudeModel(
-                project_id=vertex_project,
-                location=vertex_location,
-                model_name=model_name,
-            )
-        elif "gemini" in model_name.lower():
-            logger.info(
-                f"Creating Vertex AI Gemini model: {model_name} "
-                f"(project: {vertex_project}, location: {vertex_location})"
-            )
-            return VertexAIGeminiModel(
-                project_id=vertex_project,
-                location=vertex_location,
-                model_name=model_name,
-            )
-        else:
-            raise ValueError(f"Unknown Vertex AI model type: {model_name}")
-
-    elif provider == "vertex_ai_model_garden":
-        # Open-source models (Llama, Mistral, etc.) hosted on Vertex AI Model Garden
-        # Uses litellm's vertex_ai/ prefix for routing and Google ADC for auth
-        vertex_project = model_config.get("vertex_project")
-        vertex_location = model_config.get("vertex_location")
-
-        if not vertex_project or vertex_project == "your-gcp-project-id":
-            raise ValueError(
-                "VERTEX_PROJECT not configured. Please set VERTEX_PROJECT "
-                "environment variable in your .env file"
-            )
-
-        model_name = f"vertex_ai/{model_config['model_name']}"
-
-        logger.info(
-            f"Creating Vertex AI Model Garden model: {model_name} "
-            f"(project: {vertex_project}, location: {vertex_location})"
-        )
-
-        return LiteLLMLanguageModel(
-            model_name=model_name,
-            vertex_project=vertex_project,
-            vertex_location=vertex_location,
-        )
-
-    else:
-        # All other models use OpenAI-compatible endpoints
-        # This includes: OpenAI, OpenRouter (Claude, Gemini), Together AI (open-source), vLLM
-        api_key = get_api_key(model_id)
-
-        logger.info(f"Creating OpenAI-compatible model: {model_config['model_name']} via {model_config['base_url']}")
-
-        return OpenAICompatibleLanguageModel(
-            endpoint=model_config["base_url"],
-            api_key=api_key,
-            model_name=model_config["model_name"],
-            system_prompt=system_prompt,
-        )
-
-
-async def run_baseline(
-    lm: AbstractLanguageModel,
-    question: str,
-    enable_tools: bool = False
-) -> tuple[str, int, int, int, list[ToolCall] | None]:
-    """
-    Run baseline inference (single completion, no ITS).
-
-    Returns:
-        (answer, latency_ms, input_tokens, output_tokens, tool_calls)
-    """
-    start_time = time.time()
-
-    messages = [ChatMessage(role="user", content=question)]
-
-    # Prepare tools if enabled
-    tools = get_tool_schemas() if enable_tools else None
-    tool_choice = "auto" if enable_tools else None
-
-    # Get response and try to capture usage information
-    input_tokens = 0
-    output_tokens = 0
-
-    # For OpenAI-compatible models, try to get usage directly via litellm
-    if isinstance(lm, OpenAICompatibleLanguageModel):
-        try:
-            request_data = lm._prepare_request_data(
-                messages,
-                stop=None,
-                max_tokens=None,
-                temperature=None,
-                include_stop_str_in_output=None,
-                tools=tools,
-                tool_choice=tool_choice,
-            )
-            # _prepare_request_data is designed for aiohttp (auth via headers),
-            # but litellm needs api_key and api_base as explicit kwargs.
-            # Without these, litellm can't authenticate with non-OpenAI
-            # providers (e.g. OpenRouter) and the call fails silently.
-            request_data["api_key"] = lm.api_key
-            request_data["api_base"] = lm.endpoint
-            # litellm needs a provider prefix to route non-OpenAI models correctly.
-            # Without it, models like "meta-llama/llama-3.2-3b-instruct" fail with
-            # "LLM Provider NOT provided". Detect OpenRouter by base_url.
-            if "openrouter.ai" in lm.endpoint and not request_data.get("model", "").startswith("openrouter/"):
-                request_data["model"] = "openrouter/" + request_data["model"]
-            full_response = await litellm.acompletion(**request_data)
-
-            # Extract usage from full response
-            if hasattr(full_response, 'usage') and full_response.usage:
-                input_tokens = getattr(full_response.usage, 'prompt_tokens', 0)
-                output_tokens = getattr(full_response.usage, 'completion_tokens', 0)
-
-            # Extract message
-            response = full_response.choices[0].message.dict()
-        except Exception as e:
-            logger.warning(f"Could not capture token usage: {e}")
-            response = await lm.agenerate(messages)
-    else:
-        # For Vertex AI or other models, use standard interface
-        response = await lm.agenerate(messages)
-
-    answer = extract_content_from_lm_response(response)
-
-    # Extract and execute tool calls if present
-    tool_calls_list = None
-    if enable_tools and "tool_calls" in response and response["tool_calls"]:
-        tool_calls_list = []
-        for tc in response["tool_calls"]:
-            tool_name = tc.get("function", {}).get("name", "")
-            tool_args = parse_tool_args(tc.get("function", {}).get("arguments", {}))
-            tool_result = execute_tool(tool_name, tool_args)
-
-            tool_calls_list.append(ToolCall(
-                name=tool_name,
-                arguments=tool_args,
-                result=tool_result
-            ))
-
-            if tool_result:
-                answer += f"\n\n**Tool Used:** {tool_name}\n**Result:** {tool_result}"
-
-    latency_ms = int((time.time() - start_time) * 1000)
-
-    return answer, latency_ms, input_tokens, output_tokens, tool_calls_list
-
-
-def _build_pf_trace(result: ParticleFilteringResult) -> ParticleFilteringTrace:
-    """Build a ParticleFilteringTrace from a ParticleFilteringResult."""
-    log_w = result.log_weights_lst
-    # Normalize log-weights to probabilities
-    max_lw = max(log_w) if log_w else 0.0
-    exp_w = [np.exp(lw - max_lw) for lw in log_w]
-    sum_w = sum(exp_w)
-    normalized = [w / sum_w for w in exp_w] if sum_w > 0 else [1.0 / len(log_w)] * len(log_w)
-
-    candidates = []
-    for i, resp in enumerate(result.responses):
-        content = extract_content_from_lm_response(resp)
-        candidates.append(CandidateResponse(
-            index=i,
-            content=content,
-            is_selected=(i == result.selected_index),
-        ))
-
-    return ParticleFilteringTrace(
-        candidates=candidates,
-        log_weights=[round(w, 4) for w in log_w],
-        normalized_weights=[round(w, 4) for w in normalized],
-        steps_used=result.steps_used_lst,
-    )
-
-
-def build_trace(algorithm: str, result, tool_vote: str | None = None) -> dict | None:
-    """Convert an algorithm Result object into a serializable trace dict."""
-    try:
-        logger.debug(f"Building trace for algorithm={algorithm}, tool_vote={tool_vote}, result type={type(result)}")
-        if isinstance(result, SelfConsistencyResult):
-            candidates = []
-            for i, resp in enumerate(result.responses):
-                content = extract_content_from_lm_response(resp)
-                # Extract tool calls if present
-                tool_calls_data = None
-                if "tool_calls" in resp and resp["tool_calls"]:
-                    tool_calls_data = []
-                    for tc in resp["tool_calls"]:
-                        tool_name = tc.get("function", {}).get("name", "unknown")
-                        tool_args = parse_tool_args(tc.get("function", {}).get("arguments", {}))
-                        tool_calls_data.append(ToolCall(
-                            name=tool_name,
-                            arguments=tool_args,
-                            result=None,
-                        ))
-
-                candidates.append(CandidateResponse(
-                    index=i,
-                    content=content,
-                    is_selected=(i == result.selected_index),
-                    tool_calls=tool_calls_data,
-                ))
-
-            vote_counts = {str(k): v for k, v in result.response_counts.items()}
-
-            # Build tool voting trace if tool voting was used
-            tool_voting_trace = None
-            if tool_vote and any(c.tool_calls for c in candidates):
-                # Extract tool voting statistics
-                tool_counts = {}
-                total_tool_calls = 0
-                for candidate in candidates:
-                    if candidate.tool_calls:
-                        total_tool_calls += 1
-                        for tc in candidate.tool_calls:
-                            key = tc.name if tool_vote == "tool_name" else str(tc.arguments)
-                            tool_counts[key] = tool_counts.get(key, 0) + 1
-
-                # Find winning tool (most votes)
-                winning_tool = max(tool_counts, key=tool_counts.get) if tool_counts else "none"
-
-                tool_voting_trace = ToolVotingTrace(
-                    tool_vote_type=tool_vote,
-                    tool_counts=tool_counts,
-                    winning_tool=winning_tool,
-                    total_tool_calls=total_tool_calls,
-                )
-
-            trace = SelfConsistencyTrace(
-                candidates=candidates,
-                vote_counts=vote_counts,
-                total_votes=sum(result.response_counts.values()),
-                tool_voting=tool_voting_trace,
-            )
-            return trace.model_dump()
-
-        elif isinstance(result, BestOfNResult):
-            candidates = []
-            for i, resp in enumerate(result.responses):
-                content = extract_content_from_lm_response(resp)
-                candidates.append(CandidateResponse(
-                    index=i,
-                    content=content,
-                    is_selected=(i == result.selected_index),
-                ))
-            trace = BestOfNTrace(
-                candidates=candidates,
-                scores=[round(s, 4) for s in result.scores],
-                max_score=round(max(result.scores), 4),
-                min_score=round(min(result.scores), 4),
-            )
-            return trace.model_dump()
-
-        elif isinstance(result, BeamSearchResult):
-            candidates = []
-            for i, resp in enumerate(result.responses):
-                content = extract_content_from_lm_response(resp)
-                candidates.append(CandidateResponse(
-                    index=i,
-                    content=content,
-                    is_selected=(i == result.selected_index),
-                ))
-            trace = BeamSearchTrace(
-                candidates=candidates,
-                scores=[round(s, 4) for s in result.scores],
-                steps_used=result.steps_used,
-            )
-            return trace.model_dump()
-
-        elif isinstance(result, ParticleGibbsResult):
-            iterations = []
-            num_iterations = len(result.responses_lst)
-            for it_idx in range(num_iterations):
-                # Build a ParticleFilteringResult-like object for each iteration
-                it_result = ParticleFilteringResult(
-                    responses=result.responses_lst[it_idx],
-                    log_weights_lst=result.log_weights_lst[it_idx],
-                    selected_index=result.selected_index if it_idx == num_iterations - 1 else 0,
-                    steps_used_lst=result.steps_used_lst[it_idx],
-                )
-                iterations.append(_build_pf_trace(it_result))
-            trace = ParticleGibbsTrace(
-                num_iterations=num_iterations,
-                iterations=iterations,
-            )
-            return trace.model_dump()
-
-        elif isinstance(result, ParticleFilteringResult):
-            trace = _build_pf_trace(result)
-            return trace.model_dump()
-
-        else:
-            logger.warning(f"Unknown result type for trace: {type(result)}")
-            return None
-
-    except Exception as e:
-        logger.error(f"Failed to build trace for {algorithm}: {e}", exc_info=True)
-        return None
-
-
-async def run_its(
-    lm: OpenAICompatibleLanguageModel,
-    question: str,
-    algorithm: str,
-    budget: int,
-    api_key: str,
-    baseline_input_tokens: int = 0,
-    baseline_output_tokens: int = 0,
-    enable_tools: bool = False,
-    tool_vote: str | None = None,
-    exclude_args: list[str] | None = None,
-    question_type: str = "general",
-    judge_criterion: str = "overall_quality",
-) -> tuple[str, int, int, int, dict | None, list[ToolCall] | None]:
-    """
-    Run ITS inference with the specified algorithm.
-
-    Args:
-        lm: Language model to use
-        question: Question to answer
-        algorithm: ITS algorithm name
-        budget: Computational budget
-        api_key: API key for judge/PRM
-        baseline_input_tokens: Baseline input token count for estimation
-        baseline_output_tokens: Baseline output token count for estimation
-        enable_tools: Whether tools are enabled
-        tool_vote: Tool voting strategy
-        exclude_args: Arguments to exclude from tool voting
-        question_type: Question type ("math", "tool_calling", or "general")
-        judge_criterion: Judge criterion for Best-of-N (built-in name or custom prompt)
-
-    Returns:
-        (answer, latency_ms, input_tokens, output_tokens, trace, tool_calls)
-    """
-    start_time = time.time()
-
-    # Prepare tools if enabled
-    tools = get_tool_schemas() if enable_tools else None
-    tool_choice = "auto" if enable_tools else None
-
-    # Create algorithm instance
-    if algorithm == "best_of_n":
-        # Resolve judge criterion: built-in or custom
-        built_in_criteria = {"overall_quality", "multi_step_tool_judge"}
-        if judge_criterion in built_in_criteria:
-            criterion_to_use = judge_criterion
-        else:
-            # Custom criterion — register with CriterionRegistry
-            from reward_hub.llm_judge.prompts import Criterion, CriterionRegistry
-            criterion_name = f"custom_{hash(judge_criterion) & 0xFFFFFFFF:08x}"
-            logger.info(f"Registering custom judge criterion as: {criterion_name}")
-            custom_criterion = Criterion(
-                name=criterion_name,
-                content=judge_criterion,
-                description="Custom evaluation criterion",
-            )
-            CriterionRegistry.register(custom_criterion)
-            criterion_to_use = criterion_name
-
-        # Use LLM judge for Best-of-N
-        judge = LLMJudgeRewardModel(
-            model=DEFAULT_JUDGE_MODEL,
-            criterion=criterion_to_use,
-            judge_type="pointwise",
-            api_key=api_key,
-            enable_judge_logging=False,
-        )
-        alg = BestOfN(judge)
-
-    elif algorithm == "self_consistency":
-        # Configure Self-Consistency based on question type
-        if question_type == "tool_calling":
-            # Tool consensus: vote on tool selection
-            alg = SelfConsistency(
-                consistency_space_projection_func=None,
-                tool_vote=tool_vote or "tool_name",
-                exclude_args=exclude_args or []
-            )
-        elif question_type == "math":
-            # Math: extract boxed answers for voting
-            projection_func = create_math_projection_function()
-            alg = SelfConsistency(
-                consistency_space_projection_func=projection_func,
-                tool_vote=None,
-                exclude_args=[]
-            )
-        else:
-            # General: exact text matching (current behavior)
-            alg = SelfConsistency(
-                consistency_space_projection_func=None,
-                tool_vote=None,
-                exclude_args=[]
-            )
-
-    elif algorithm in ["beam_search", "particle_filtering", "entropic_particle_filtering", "particle_gibbs"]:
-        # Process-based algorithms need StepGeneration and Process Reward Model
-
-        # Create StepGeneration with step-by-step reasoning
-        # Using "\n\n" as step delimiter for reasoning steps
-        step_gen = StepGeneration(
-            max_steps=DEFAULT_STEP_GEN_MAX_STEPS,
-            step_token=DEFAULT_STEP_GEN_TOKEN,
-            stop_token=None,
-            temperature=DEFAULT_STEP_GEN_TEMPERATURE,
-            include_stop_str_in_output=False,
-        )
-
-        prm = LLMProcessRewardModel(
-            model=DEFAULT_PRM_MODEL,
-            api_key=api_key,
-            temperature=DEFAULT_PRM_TEMPERATURE,
-        )
-
-        if algorithm == "beam_search":
-            beam_width = max(BEAM_WIDTH_MIN, min(BEAM_WIDTH_MAX, budget // 2))
-            adjusted_budget = max(beam_width, (budget // beam_width) * beam_width)
-            alg = BeamSearch(sg=step_gen, prm=prm, beam_width=beam_width)
-            budget = adjusted_budget
-
-        elif algorithm == "particle_filtering":
-            # Basic particle filtering
-            alg = ParticleFiltering(
-                sg=step_gen,
-                prm=prm,
-            )
-
-        elif algorithm == "entropic_particle_filtering":
-            # Entropic particle filtering with temperature annealing
-            alg = EntropicParticleFiltering(
-                sg=step_gen,
-                prm=prm,
-            )
-
-        elif algorithm == "particle_gibbs":
-            # Particle Gibbs with multiple iterations
-            # Use smaller budget for iterations to avoid timeout
-            num_iterations = max(GIBBS_ITERATIONS_MIN, min(GIBBS_ITERATIONS_MAX, budget // GIBBS_ITERATIONS_DIVISOR))
-            alg = ParticleGibbs(
-                sg=step_gen,
-                prm=prm,
-                num_iterations=num_iterations,
-            )
-    else:
-        raise ValueError(f"Unknown algorithm: {algorithm}")
-
-    # Run inference with full result to capture trace data
-    result = await alg.ainfer(
-        lm,
-        question,
-        budget=budget,
-        return_response_only=False,
-        tools=tools,
-        tool_choice=tool_choice
-    )
-    answer = extract_content_from_lm_response(result.the_one)
-
-    # Extract and execute tool calls if present
-    tool_calls_list = None
-    if enable_tools and "tool_calls" in result.the_one and result.the_one["tool_calls"]:
-        tool_calls_list = []
-        for tc in result.the_one["tool_calls"]:
-            tool_name = tc.get("function", {}).get("name", "")
-            tool_args = parse_tool_args(tc.get("function", {}).get("arguments", {}))
-            tool_result = execute_tool(tool_name, tool_args)
-
-            tool_calls_list.append(ToolCall(
-                name=tool_name,
-                arguments=tool_args,
-                result=tool_result
-            ))
-
-            if tool_result:
-                answer += f"\n\n**Tool Used:** {tool_name}\n**Result:** {tool_result}"
-
-    latency_ms = int((time.time() - start_time) * 1000)
-
-    # Build trace from the full result
-    trace = build_trace(algorithm, result, tool_vote=tool_vote)
-
-    # --- Estimated token usage for ITS ---
-    #
-    # WHY ESTIMATED: ITS algorithms call the LLM multiple times internally
-    # (via ainfer()), but the algorithm classes do not accumulate or return
-    # per-call token usage from litellm. Until the algorithm internals are
-    # updated to surface aggregated usage data, we estimate based on:
-    #
-    #   1. The baseline's actual token counts (from a single litellm call)
-    #   2. The number of candidates the algorithm produced
-    #   3. Algorithm-specific overhead (judge/PRM scoring calls)
-    #
-    # These estimates are marked with tokens_estimated=True in the response
-    # so the frontend can label them accordingly.
-    #
-    # WHAT IS ACTUAL vs ESTIMATED for ITS results:
-    #   - latency_ms:    ACTUAL (wall-clock around the full ainfer() call)
-    #   - input_tokens:  ESTIMATED (see formula below)
-    #   - output_tokens: ESTIMATED (see formula below)
-    #   - cost_usd:      ESTIMATED (derived from estimated tokens x pricing)
-    #
-    num_candidates = len(result.responses) if hasattr(result, 'responses') else budget
-
-    if baseline_input_tokens > 0 and baseline_output_tokens > 0:
-        # Base: assume each candidate uses roughly the same tokens as baseline
-        estimated_input = baseline_input_tokens * num_candidates
-        estimated_output = baseline_output_tokens * num_candidates
-
-        if algorithm == "best_of_n":
-            # Best-of-N also makes one LLM judge call per candidate to score
-            # it. The judge sees the candidate response (~baseline_output_tokens)
-            # plus a scoring prompt (~200 tokens), and produces a short score
-            # output (~50 tokens).
-            judge_input_per_call = baseline_output_tokens + 200
-            judge_output_per_call = 50
-            estimated_input += judge_input_per_call * num_candidates
-            estimated_output += judge_output_per_call * num_candidates
-        elif algorithm in ["beam_search", "particle_filtering", "entropic_particle_filtering", "particle_gibbs"]:
-            # Process-based algorithms make PRM scoring calls at each
-            # reasoning step. We use the actual step count from the result
-            # when available, otherwise fall back to budget * 4 as a rough
-            # average. Each PRM call uses ~150 input and ~30 output tokens.
-            if hasattr(result, 'steps_used_lst'):
-                prm_calls = sum(result.steps_used_lst)
-            else:
-                prm_calls = budget * 4
-            estimated_input += 150 * prm_calls
-            estimated_output += 30 * prm_calls
-    else:
-        estimated_input = 0
-        estimated_output = 0
-
-    return answer, latency_ms, estimated_input, estimated_output, trace, tool_calls_list
 
 
 @app.post("/compare", response_model=CompareResponse)
